+++ CHANGES.rst
@@ -13,8 +13,8 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 
 - Update to newest compatible versions of dependencies.
 
-- Use intermediate ``str`` representation for complex response data (i.e. not
-  ``str`` or ``bytes`` like)
+- Use intermediate ``str`` representation for non-bytelike response data unless
+  indicated differently by the content type.
   (`#1006 <https://github.com/zopefoundation/Zope/issues/1006>`_)